requirements.txt
praw
yfinance
google-genai
flask
flask-cors
google-generativeai
gunicorn